order pushes to firebase and heads to confirmation page

+++ RoseEats/Models/MenuItem.swift
@@ -12,7 +12,7 @@ import Firebase
 class MenuItem {
     var id: String!
     var Name: String!
-    var Price: Int!
+    var Price: Float!
     var WhereServed: [String]!
     var ImageName: String!
     
@@ -20,7 +20,7 @@ class MenuItem {
         self.id = documentSnapShot.documentID
         let data = documentSnapShot.data()!
         self.Name = (data["Name"] as! String)
-        self.Price = data["Location"] as? Int
+        self.Price = data["Location"] as? Float
         self.WhereServed = data["WhereServed"] as? [String]
         self.ImageName = data["ImageName"] as? String
     }
